Adding Node Class, adding toString Method to AdjMat class, Finishing MST Generation Loop
The Node class was added, which is used for storing the structure of the
MST. This will be used later to enable quickly determining the greatest
edge weight in the path between two nodes along the MST. Doing this
quickly is essential for being able to rapidly add edges to the graph
without changing its MST.

The toString method of the AdjMat class converts the adjacency Matrix
into a square matrix representation, with Unicode multipart square
brackets on either side for presentation. This method currently causes
broken formatting when numbers ar allowed to be over 99. I'll probably
make another method for padding out numbers so that this doesn't happen,
but for now it's what we've got.

Also, the toString will obviously break if the rows of the graph are
wider than the available line width, since each row would wrap around in
an unsightly fashion. Nothing I can do about this right now, sadly...