Modifying ReadMe and Finishing Graph Generation (No output yet)
ReadMe was changed to reflect changes in the intention of the algorithm.
Originally, I wanted to make all edge weights in the graph unique.
However, this would incur an additional n^2 cost which I wanted to
avoid.

Since the algorithm is already n^2, I might revert this, but I would
have to do performance testing. in any case the result should still have
exactly 1 MST.

Also, the edge-generating part of the genGraph function was essentially
finished. I also wrote in a new helper method for finding the heaviest
node in the path between nodes.