Removing Adjacency Matrix console.logs, adding console.time for various parts of the genGraph method
I removed some console.log statements which I worried were causing
performance issues, and added some console.time() statements for help in
profiling. I found that the seemingly slowest step is building the
output strings. That said, the timing for this is reported long before
the output boxes are actually re-painted on the interface. This
indicates to me that performance is probably constrained by these two
steps moreso than by the algorithm used.

I'm fairly certain this method for building the graph is O(n^2 log n)
for n = the number of vertices on the graph (noting that the graph is
saturated).

This makes the algorithm equivalent in complexity to simply generating a
random graph and solving its MST (since MST finding algorithms are O(e
log n), where e is the number of edges, which is ~n^2 for a saturated
graph). However, we could do better by using another AdjMat object and
storing all of the worst path lengths. This would have O(n^2) time, but
would double memory usage at best.

Considering Chrome runs out of memory long before running time gets
ridiculous, and most of the time is spent on painting the result to the
screen anyways, I don't think trading more memory for better big O is
worth it for this particular program.