Adding Output Description
Added a description for the format of the MST output, since people were
getting confused over why the order of the output arcs and vertices was
different from their output.

I also made the background slightly blue-er to compensate for the fact
that the gray looked kinda greenish on my display. It's a very slight
change, so it's pure aesthetics.

Oh, and I removed a commented-out block in the expanding div. No reason
to have that in for now, since formatting options will be a later
feature.