Adding answer-checking feature
When the result is presented to the user, they can now print their own result and check it against the given one automagically. 

The checking function does not care about the order of arcs or vertices in the user input, instead relying on the inherent order of the script's output to do a quick O(n) match between the input and the known output.

I also made some CSS and formatting changes which prevent the MST output from getting pushed under the graph output. By my testing this should work under most browser sizes and zoom settings, but I will accept further feedback on the matter.